Add support for downloading and handling SDL native libs
This commit introduces changes to automate the download and setup of native SDL libraries, specifically for Windows x64. It updates the Makefile with a new `native` target for library handling and adjusts the GitHub workflow to include SDL artifact downloads. Additionally, the `.gitignore` file was updated to exclude the `downloads` directory.